+++ src/main/java/com/groute/groute_server/common/config/WebMvcConfig.java
@@ -6,7 +6,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
-import com.groute.groute_server.auth.resolver.CurrentUserArgumentResolver;
+import com.groute.groute_server.common.resolver.CurrentUserArgumentResolver;
 
 import lombok.RequiredArgsConstructor;
