+++ src/main/java/com/groute/groute_server/common/jwt/JwtProperties.java
@@ -1,4 +1,4 @@
-package com.groute.groute_server.auth.jwt;
+package com.groute.groute_server.common.jwt;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
